refactor(#12): 스타일 컴포넌트 common 폴더로 이동
